Scripts I made for my first demo thing
Putting the Moveable_Object.cs script on an object with a Rigidbody will make it moveable. Clicking the, "Active" checkmark in the script in the inspector will make that object controllable at the start of the game. Possession Distance should be a little further than the maximum length an object reaches from its center. Changing torque multiplier will increase how fast the object can spin, combined with changing its weight in Rigidbody.

The camera has to have the tag, "MainCamera" and the StolenFollowCamera.cs script, and be under an empty game object to hold the camera to work. Rotate speed 5 and Height Offset -5 make the most sense to me. I was thinking about adding functionality to change the height offset and distance from the object on a per-object basis so it can change with bigger or smaller objects, but idk if that's something you guys would want. Just lmk if you have any questions or I should change something